refactor: new CollapsibleHeaderScrollView component
