+++ src/features/dashboard/CollapsibleHeaderScrollView.tsx
@@ -1,0 +1,121 @@
+import React, { useState } from 'react';
+import { View, Animated, ScrollView, StyleSheet, Dimensions } from 'react-native';
+
+import { colors } from '@theme';
+
+const { width: SCREEN_WIDTH } = Dimensions.get('screen');
+
+type CollapsibleHeaderHeightConfig = {
+  compact: number;
+  expanded: number;
+};
+
+interface CollapsibleHeaderScrollViewProps {
+  compactHeader: React.ReactNode;
+  expandedHeader: React.ReactNode;
+  config: CollapsibleHeaderHeightConfig;
+}
+
+export const CollapsibleHeaderScrollView: React.FC<CollapsibleHeaderScrollViewProps> = ({
+  compactHeader,
+  expandedHeader,
+  config,
+  children,
+}) => {
+  const [scrollY, _] = useState<Animated.Value>(new Animated.Value(0));
+
+  const headerHeight = scrollY.interpolate({
+    inputRange: [0, config.expanded - config.compact],
+    outputRange: [config.expanded, config.compact],
+    extrapolate: 'clamp',
+  });
+
+  // Fade in Compact header as user scroll down
+  const compactHeaderOpacity = scrollY.interpolate({
+    inputRange: [config.compact, config.expanded - config.compact],
+    outputRange: [0, 1],
+    extrapolate: 'clamp',
+  });
+
+  // Fade out Expanded header as user scroll down
+  const expanedHeaderOpacity = scrollY.interpolate({
+    inputRange: [0, config.expanded - config.compact - 75],
+    outputRange: [1, 0],
+    extrapolate: 'clamp',
+  });
+
+  // Slide up Compact header as user scroll down
+  const compactHeaderY = scrollY.interpolate({
+    inputRange: [0, config.expanded - config.compact],
+    outputRange: [50, 0],
+    extrapolate: 'clamp',
+  });
+
+  // Slide up Expanded header as user scroll down
+  const expanedHeaderY = scrollY.interpolate({
+    inputRange: [0, config.expanded - config.compact],
+    outputRange: [0, -25],
+    extrapolate: 'clamp',
+  });
+
+  return (
+    <View style={styles.container}>
+      <Animated.View style={[styles.header, { height: headerHeight }]}>
+        <Animated.View
+          style={{
+            paddingTop: compactHeaderY,
+            opacity: compactHeaderOpacity,
+          }}>
+          {compactHeader}
+        </Animated.View>
+        <Animated.View
+          style={[
+            {
+              opacity: expanedHeaderOpacity,
+              top: expanedHeaderY,
+            },
+            styles.expandedHeaderContainer,
+          ]}>
+          {expandedHeader}
+        </Animated.View>
+      </Animated.View>
+      <ScrollView
+        contentContainerStyle={{ paddingTop: config.expanded }}
+        scrollIndicatorInsets={{
+          top: config.expanded - 40,
+        }}
+        onScroll={Animated.event([
+          {
+            nativeEvent: {
+              contentOffset: {
+                y: scrollY,
+              },
+            },
+          },
+        ])}
+        scrollEventThrottle={16}>
+        {children}
+      </ScrollView>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#eaeaea',
+  },
+  header: {
+    backgroundColor: colors.predict,
+    position: 'absolute',
+    width: SCREEN_WIDTH,
+    top: 0,
+    left: 0,
+    zIndex: 9999,
+    overflow: 'hidden',
+  },
+  expandedHeaderContainer: {
+    position: 'absolute',
+    width: '100%',
+  },
+});
